Split rotation status bar into widgets
